- Major Re-org of roles code, plus add support for Active Directory - Minor performance improvements
